backend/requirements.txt
anyio==3.6.2
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
cloudevents==1.6.2
colorama==0.4.6
dataclasses==0.6
deprecation==2.1.0
Django==4.1.3
docutils==0.19
dotty-dict==1.3.1
Flask==2.2.2
functions-framework==3.2.1
gitdb==4.0.9
GitPython==3.1.29
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.0
google-cloud-language==2.6.1
googleapis-common-protos==1.56.4
gotrue==0.5.4
grpcio==1.50.0
grpcio-status==1.50.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
idna==3.4
importlib-metadata==5.0.0
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.11.0
MarkupSafe==2.1.1
more-itertools==9.0.0
oauthlib==3.2.2
packaging==21.3
pkginfo==1.8.3
postgrest-py==0.10.3
proto-plus==1.22.1
protobuf==4.21.10
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-gitlab==3.11.0
python-semantic-release==7.32.1
readme-renderer==37.3
realtime==0.0.5
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rsa==4.9
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlparse==0.4.3
storage3==0.3.5
supabase==0.7.1
supafunc==0.2.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
twine==3.8.0
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
uuid==1.30
watchdog==2.1.9
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
zipp==3.10.0
